+++ graphbrain/agents/actors.py
@@ -5,8 +5,8 @@ from graphbrain.agents.agent import Agent
 
 
 class Actors(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.actor_counter = None
 
     def name(self):