+++ graphbrain/agents/claim_actors.py
@@ -3,8 +3,8 @@ from graphbrain.agents.agent import Agent
 
 
 class ClaimActors(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.search_pattern = ('claim/p/.', '*', '*', '*')
 
     def name(self):