+++ graphbrain/agents/claims.py
@@ -31,8 +31,8 @@ def replace_subject(edge, new_subject):
 
 
 class Claims(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.actors = None
         self.female = None
         self.group = None