+++ graphbrain/agents/corefs_names.py
@@ -32,8 +32,8 @@ def clique_number(edge, cliques):
 
 
 class CorefsNames(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.corefs = 0
         self.seeds = None