+++ graphbrain/agents/corefs_onto.py
@@ -5,8 +5,8 @@ from graphbrain.agents.agent import Agent
 
 
 class CorefsOnto(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.corefs = 0
 
     def name(self):