+++ graphbrain/agents/corefs_unidecode.py
@@ -18,8 +18,8 @@ def unidecode_edge(edge):
 
 
 class CorefsUnidecode(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
         self.corefs = 0
 
     def name(self):