+++ graphbrain/agents/taxonomy.py
@@ -3,8 +3,8 @@ from graphbrain.agents.agent import Agent
 
 
 class Taxonomy(Agent):
-    def __init__(self, hg):
-        super().__init__(hg)
+    def __init__(self, hg, lang, sequence=None):
+        super().__init__(hg, lang, sequence)
 
     def name(self):
         return 'taxonomy'