+++ graphbrain/commands/run.py
@@ -4,7 +4,7 @@ from graphbrain.agents import create_agent
 
 def run(args):
     hg = hgraph(args.hg)
-    agent = create_agent(args.agent, hg)
+    agent = create_agent(args.agent, hg, args.lang, args.sequence)
     if agent is None:
         print('ERROR: unknown agent: {}'.format(args.agent))
     else: